+++ index.js
@@ -1,10 +1,10 @@
 'use strict';
 
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
 Object.defineProperty(exports, "__esModule", {
     value: true
 });
+
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
 var _react = require('react');
 
@@ -41,7 +41,7 @@ var InlineEdit = function (_React$Component) {
         }
 
         return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(InlineEdit)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.state = {
-            editing: false,
+            editing: _this.props.editing,
             text: _this.props.text,
             minLength: _this.props.minLength,
             maxLength: _this.props.maxLength
@@ -94,8 +94,17 @@ var InlineEdit = function (_React$Component) {
     }, {
         key: 'componentWillReceiveProps',
         value: function componentWillReceiveProps(nextProps) {
-            if (nextProps.text !== this.props.text) {
-                this.setState({ text: nextProps.text });
+            var isTextChanged = nextProps.text !== this.props.text;
+            var isEditingChanged = nextProps.editing !== this.props.editing;
+            var nextState = {};
+            if (isTextChanged) {
+                nextState.text = nextProps.text;
+            }
+            if (isEditingChanged) {
+                nextState.editing = nextProps.editing;
+            }
+            if (isTextChanged || isEditingChanged) {
+                this.setState(nextState);
             }
         }
     }, {
@@ -112,10 +121,19 @@ var InlineEdit = function (_React$Component) {
     }, {
         key: 'render',
         value: function render() {
-            if (!this.state.editing) {
+            if (this.props.isDisabled) {
                 var Element = this.props.element || this.props.staticElement;
                 return _react2.default.createElement(
                     Element,
+                    {
+                        className: this.props.className,
+                        style: this.props.style },
+                    this.state.text || this.props.placeholder
+                );
+            } else if (!this.state.editing) {
+                var _Element = this.props.element || this.props.staticElement;
+                return _react2.default.createElement(
+                    _Element,
                     {
                         className: this.props.className,
                         onClick: this.startEditing,
@@ -124,8 +142,8 @@ var InlineEdit = function (_React$Component) {
                     this.state.text || this.props.placeholder
                 );
             } else {
-                var Element = this.props.element || this.props.editingElement;
-                return _react2.default.createElement(Element, {
+                var _Element2 = this.props.element || this.props.editingElement;
+                return _react2.default.createElement(_Element2, {
                     onClick: this.clickWhenEditing,
                     onKeyDown: this.keyDown,
                     onBlur: this.finishEditing,
@@ -156,13 +174,17 @@ InlineEdit.propTypes = {
     style: _react2.default.PropTypes.object,
     editingElement: _react2.default.PropTypes.string,
     staticElement: _react2.default.PropTypes.string,
-    tabIndex: _react2.default.PropTypes.number
+    tabIndex: _react2.default.PropTypes.number,
+    isDisabled: _react2.default.PropTypes.bool,
+    editing: _react2.default.PropTypes.bool
 };
 InlineEdit.defaultProps = {
     minLength: 1,
     maxLength: 256,
     editingElement: 'input',
     staticElement: 'span',
-    tabIndex: 0
+    tabIndex: 0,
+    isDisabled: false,
+    editing: false
 };
 exports.default = InlineEdit;